+++ frontend/src/components/course-search/CourseRow.module.css
@@ -141,35 +141,6 @@
             width: 5.5ex;
         }
 
-        & .courseNumberContainer {
-            display: flex;
-            min-width: 8.5em;
-            justify-content: space-between;
-        }
-
-        & .courseNumber {
-            margin-right: 0.5em;
-            font-weight: 700;
-            font-size: 0.875em;
-            opacity: 0.875;
-            text-align: left;
-            cursor: pointer;
-            padding: 0.125rem 0.25rem;
-            border-radius: 0.25rem;
-            transition:
-                background-color 200ms,
-                opacity 200ms;
-
-            &:hover {
-                opacity: 1;
-                background-color: #0001;
-            }
-
-            &:active {
-                background-color: #0002;
-            }
-        }
-
         & .affiliation {
             width: 4.5ex;
             text-align: right;
